refactor: move UpstreamPool out of mcp/ to fix api->mcp dependency
The layer contract forbids api -> mcp dependencies. UpstreamPool was in
mcp/upstream/ but referenced by api/state.rs and api/router.rs.

- Move pool.rs and types.rs to dispatch/upstream/ (shared layer)
- mcp/upstream.rs becomes a thin re-export for backward compat
- build_mcp_service() and allowed_hosts_from_env() move from
  api/router.rs to cli/serve.rs (MCP construction is serve's job)
- build_router_with_bearer() accepts an optional pre-built MCP router
  instead of constructing LabMcpServer internally
- All references updated to use crate::dispatch::upstream
