+++ crates/lab/src/dispatch/upstream.rs
@@ -1,0 +1,15 @@
+//! Upstream MCP server proxy — shared types and connection pool.
+//!
+//! Lives in `dispatch/` (not `mcp/`) because both the MCP and API surfaces
+//! need access to `UpstreamPool`. The layer contract forbids `api -> mcp`
+//! dependencies, so shared types must live in the dispatch layer.
+//
+// Many items in pool and types are not yet called from outside the module
+// (discovery, resource proxying, circuit breaker probing). They are exercised
+// by tests and will be fully wired when `lab serve` gains `[[upstream]]` config
+// support. The blanket allow prevents false-positive warnings on partially
+// wired public APIs.
+#[allow(dead_code)]
+pub mod pool;
+#[allow(dead_code)]
+pub mod types;
