+++ crates/lab/src/mcp/upstream.rs
@@ -1,13 +1,8 @@
-//! Upstream MCP server proxy — connects to external MCP servers and
-//! exposes their tools through the lab gateway.
-//
-// Many items in pool and types are not yet called from outside the module
-// (discovery, resource proxying, circuit breaker probing). They are exercised
-// by tests and will be fully wired when `lab serve` gains `[[upstream]]` config
-// support. The blanket allow prevents false-positive warnings on partially
-// wired public APIs.
-#[allow(dead_code)]
-pub mod pool;
-#[allow(dead_code)]
-pub mod types;
+//! Re-export upstream types from the shared dispatch layer.
+//!
+//! The canonical implementation lives in `crate::dispatch::upstream`.
+//! This re-export exists so that existing `crate::mcp::upstream::*`
+//! references continue to resolve without a full codebase rename.
 
+#[allow(unused_imports)]
+pub use crate::dispatch::upstream::{pool, types};
